requirements.txt
ossapi
python-dotenv